add and update channel-specific devices
In the FHEM module, a dual actor is represented as three devices.
One main device (e.g. 4376bb) and one for each channel with appended
channel number (e.g. 4376bb01 and 4376bb02).

For consistency, this should be also done for pyduofern, but
was only partly implemented, yet. In particular, these channel-specific
devices where not added as own devices and thus could not be addressed,
leading to exceptions.

This commit adds all three devices and also sets the state
of the respective device.